Implement setMetadata

+++ packages/corporations/src/systems/CorporationsSystemErrors.sol
@@ -8,6 +8,6 @@ interface CorporationsSystemErrors {
   error CorporationsSystem_CorpNotClaimed();
   error CorporationsSystem_IsAlreadyCeo(uint256 corpId, uint256 ceoId);
   error CorporationsSystem_Unauthorized(uint256 corpId, uint256 callerId);
-  error CorporationsSystem_InvalidStringLength(string value, uint8 minLength, uint8 maxLength);
+  error CorporationsSystem_InvalidStringLength(string value, uint16 minLength, uint16 maxLength);
   error CorporationsSystem_InvalidTickerFormat(bytes8 ticker);
 }
